+++ forgerock-auth-ui/src/main/java/org/forgerock/android/auth/ui/callback/CallbackController.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 ForgeRock. All rights reserved.
+ * Copyright (c) 2020 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -28,6 +28,13 @@ public interface CallbackController {
      * @param callback The Callback
      */
     void onDataCollected(Callback callback);
+
+    /**
+     * Cancel the current authentication
+     *
+     * @param e Exception when cancelling the authentication process.
+     */
+    void cancel(Exception e);
 
 
 }