+++ forgerock-auth-ui/src/main/java/org/forgerock/android/auth/ui/callback/DeviceProfileCallbackFragment.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 ForgeRock. All rights reserved.
+ * Copyright (c) 2020 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -8,34 +8,77 @@
 package org.forgerock.android.auth.ui.callback;
 
 
+import android.content.pm.PackageManager;
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
 
 import org.forgerock.android.auth.FRListener;
-import org.forgerock.android.auth.callback.DeviceAttributeCallback;
+import org.forgerock.android.auth.callback.DeviceProfileCallback;
 import org.forgerock.android.auth.ui.R;
+
+import static android.Manifest.permission.ACCESS_FINE_LOCATION;
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class DeviceProfileCallbackFragment extends  CallbackFragment<DeviceAttributeCallback> {
+public class DeviceProfileCallbackFragment extends CallbackFragment<DeviceProfileCallback> {
 
+    private TextView message;
+    public static final int LOCATION_REQUEST_CODE = 100;
 
     public DeviceProfileCallbackFragment() {
         // Required empty public constructor
     }
 
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        //No matter permission is granted or not, we proceed to the next node.
+        proceed();
+    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
         View view = inflater.inflate(R.layout.fragment_device_profile_callback, container, false);
+        message = view.findViewById(R.id.message);
+        message.setText(callback.getMessage());
+
+        if (callback.isLocation()) {
+            if (ContextCompat.checkSelfPermission(getContext(), ACCESS_FINE_LOCATION)
+                    != PackageManager.PERMISSION_GRANTED) {
+                //We don't have permission.
+                if (shouldShowRequestPermissionRationale(ACCESS_FINE_LOCATION)) {
+                    AlertDialog alertDialog = new AlertDialog.Builder(getContext(), R.style.AlertDialogTheme)
+                            .setMessage(R.string.request_location_rationale)
+                            .setPositiveButton("Proceed",
+                                    (dialog, which) -> requestLocationPermission())
+                            .setNegativeButton("Deny", (dialog, which) -> proceed())
+                            .create();
+                    alertDialog.show();
+                } else {
+                    requestLocationPermission();
+                }
+            } else {
+                // Location permission is granted, proceed.
+                proceed();
+            }
+        } else {
+            proceed();
+        }
+
+       return view;
+    }
+
+    private void proceed() {
         callback.execute(this.getContext(), new FRListener<Void>() {
             @Override
             public void onSuccess(Void result) {
@@ -44,10 +87,14 @@ public class DeviceProfileCallbackFragment extends  CallbackFragment<DeviceAttri
 
             @Override
             public void onException(Exception e) {
-                //Do Something
+                //Not likely to happen, Device Collector try best to collect.
+                cancel(e);
             }
         });
-        return view;
     }
 
+    private void requestLocationPermission() {
+        requestPermissions(new String[]{ACCESS_FINE_LOCATION}
+                , LOCATION_REQUEST_CODE);
+    }
 }