+++ forgerock-auth/src/main/java/org/forgerock/android/auth/Config.java
@@ -42,6 +42,12 @@ public class Config {
     private int timeout;
     private List<String> pins;
     private CookieJar cookieJar;
+    private String authenticateEndpoint;
+    private String authorizeEndpoint;
+    private String tokenEndpoint;
+    private String revokeEndpoint;
+    private String userinfoEndpoint;
+    private String logoutEndpoint;
 
     //SSO Token Manager
     private String accountName;
@@ -50,6 +56,7 @@ public class Config {
     //Token Manager
     private SharedPreferences sharedPreferences;
     private long cacheIntervalMillis = 0L;
+    private long cookieCacheIntervalMillis = 0L;
     private long threshold;
 
     //KeyStoreManager
@@ -61,13 +68,21 @@ public class Config {
         redirectUri = context.getString(R.string.forgerock_oauth_redirect_uri);
         scope = context.getString(R.string.forgerock_oauth_scope);
         oAuthUrl = context.getString(R.string.forgerock_oauth_url);
+        cacheIntervalMillis = context.getResources().getInteger(R.integer.forgerock_oauth_cache) * 1000;
         threshold = context.getResources().getInteger(R.integer.forgerock_oauth_threshold);
         url = context.getString(R.string.forgerock_url);
         realm = context.getString(R.string.forgerock_realm);
         timeout = context.getResources().getInteger(R.integer.forgerock_timeout);
         accountName = context.getString(R.string.forgerock_account_name);
-        cookieJar = CookieJar.NO_COOKIES;
+        cookieJar = null; // We cannot initialize default cookie jar here
+        cookieCacheIntervalMillis = context.getResources().getInteger(R.integer.forgerock_cookie_cache) * 1000;
         pins = Arrays.asList(context.getResources().getStringArray(R.array.forgerock_pins));
+        authenticateEndpoint = context.getString(R.string.forgerock_authenticate_endpoint);
+        authorizeEndpoint = context.getString(R.string.forgerock_authorize_endpoint);
+        tokenEndpoint = context.getString(R.string.forgerock_token_endpoint);
+        revokeEndpoint = context.getString(R.string.forgerock_revoke_endpoint);
+        userinfoEndpoint = context.getString(R.string.forgerock_userinfo_endpoint);
+        logoutEndpoint = context.getString(R.string.forgerock_logout_endpoint);
     }
 
     public static Config getInstance(Context context) {
@@ -94,7 +109,13 @@ public class Config {
                 .url(url)
                 .realm(realm)
                 .timeout(timeout)
-                .cookieJar(cookieJar)
+                .cookieJarSupplier(() -> getCookieJar())
+                .authenticateEndpoint(authenticateEndpoint)
+                .authorizeEndpoint(authorizeEndpoint)
+                .tokenEndpoint(tokenEndpoint)
+                .revokeEndpoint(revokeEndpoint)
+                .userInfoEndpoint(userinfoEndpoint)
+                .logoutEndpoint(logoutEndpoint)
                 .build();
     }
 
@@ -131,6 +152,13 @@ public class Config {
                 .oAuth2Client(getOAuth2Client())
                 .build();
 
+    }
+
+    private CookieJar getCookieJar() {
+        if (cookieJar == null) {
+            cookieJar = SecureCookieJar.builder().build();
+        }
+        return cookieJar;
     }
 
     @VisibleForTesting
@@ -190,6 +218,10 @@ public class Config {
 
     Long applyDefaultIfNull(Long cacheIntervalMillis) {
         return applyIfNull(cacheIntervalMillis, null, (Function<Void, Long>) var -> getCacheIntervalMillis());
+    }
+
+    Long applyDefaultCookieCacheIfNull(Long cookieCache) {
+        return applyIfNull(cookieCache, null, (Function<Void, Long>) var -> getCookieCacheIntervalMillis());
     }
 
     Long applyDefaultThresholdIfNull(Long threshold) {