+++ forgerock-core/src/main/java/org/forgerock/android/auth/ServerConfig.java
@@ -43,7 +43,17 @@ public class ServerConfig {
 
     private String host;
 
-    private CookieJar cookieJar;
+    private Supplier<CookieJar> cookieJarSupplier;
+
+    /**
+     * Server Endpoint setting, leave it empty to use default setting.
+     */
+    private String authenticateEndpoint;
+    private String authorizeEndpoint;
+    private String tokenEndpoint;
+    private String revokeEndpoint;
+    private String userInfoEndpoint;
+    private String logoutEndpoint;
 
     @lombok.Builder
     public ServerConfig(@NonNull Context context,
@@ -51,8 +61,14 @@ public class ServerConfig {
                         String realm,
                         Integer timeout,
                         TimeUnit timeUnit,
-                        CookieJar cookieJar,
-                        @Singular List<String> pins) {
+                        Supplier<CookieJar> cookieJarSupplier,
+                        @Singular List<String> pins,
+                        String authenticateEndpoint,
+                        String authorizeEndpoint,
+                        String tokenEndpoint,
+                        String revokeEndpoint,
+                        String userInfoEndpoint,
+                        String logoutEndpoint) {
 
         this.url = url;
         try {
@@ -65,7 +81,20 @@ public class ServerConfig {
         this.timeout = timeout == null ? context.getResources().getInteger(R.integer.forgerock_timeout) : timeout;
         this.timeUnit = timeUnit == null ? SECONDS : timeUnit;
         this.pins = Arrays.asList(context.getResources().getStringArray(R.array.forgerock_pins));
-        this.cookieJar = cookieJar;
+        this.cookieJarSupplier = cookieJarSupplier;
+        this.authenticateEndpoint = authenticateEndpoint;
+        this.authorizeEndpoint = authorizeEndpoint;
+        this.tokenEndpoint = tokenEndpoint;
+        this.revokeEndpoint = revokeEndpoint;
+        this.userInfoEndpoint = userInfoEndpoint;
+        this.logoutEndpoint = logoutEndpoint;
     }
 
+    CookieJar getCookieJar() {
+        if (cookieJarSupplier != null) {
+            return cookieJarSupplier.get();
+        } else {
+            return CookieJar.NO_COOKIES;
+        }
+    }
 }