+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## [1.0.1]
 
+#### Added
+- Support Set Persistent Cookie Node [SDKS-182]
+- Support Device Profile Collector Node [SDKS-293]
+- Support Metadata Callback & Stage Mapping for 6.5.2 [SDKS-305]
+- Allow server paths to be configurable [SDKS-307]
+
 #### Fixed
 - Interface alignment with other platforms and introduce FRSession [SDKS-177]
 - FRAuth.next() is now deprecated, user FRSession.authenticate instead [SDKS-177] 
 - Fix Instrument Test. [SDKS-162]
-- Support Set Persistent Cookie Node [SDKS-182]
-
 
 ## [1.0.0]
 - General Availability release for SDKs